refactor to extract Listing 13-6
